move remaining content into own folders
closes #329
