+++ projects/ngx-d3timeline/src/lib/content/content-utils.ts
@@ -1,5 +1,5 @@
 import { PositionedActivity } from '../activity/positioned-activity';
-import { TimeScale, BandScale } from '../scale-types';
+import { TimeScale, BandScale } from '../scales/scale-types';
 import { Orientation } from '../core/orientation';
 import { Point, translatePoint, pointToTransform } from '../core/point';
 import { ActivityRectangle } from '../activity-rectangle/activity-rectangle';
