+++ projects/ngx-d3timeline/src/lib/scales/scale-types.ts
