+++ projects/ngx-d3timeline/src/lib/tick-mark/resource-axis-tick-mark-renderer.ts
@@ -1,4 +1,4 @@
-import { BandScale } from '../scale-types';
+import { BandScale } from '../scales/scale-types';
 import {
   TickMarkRenderer,
   tickLabelSpacing
