+++ projects/ngx-d3timeline/src/lib/tick-mark/time-axis-tick-mark-renderer.ts
@@ -1,4 +1,4 @@
-import { TimeScale } from '../scale-types';
+import { TimeScale } from '../scales/scale-types';
 import { TickMarkRenderer, tickLabelSpacing } from './tick-mark-renderer';
 import { Orientation } from '../core/orientation';
